requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
cryptography==41.0.5
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0 